+++ package.json
@@ -36,6 +36,7 @@
 		"jest": "^29.4.3",
 		"jest-environment-jsdom": "^29.4.3",
 		"jest-transform-stub": "^2.0.0",
+		"react-router-dom": "^6.8.2",
 		"styled-components": "^5.3.6",
 		"ts-jest": "^29.0.5",
 		"typescript": "^4.9.3",
