+++ src/components/pages/Pokedex/index.ts
@@ -1,0 +1,1 @@
+export * from "./Pokedex";
