+++ src/components/pages/PokemonFinder/PokemonFinder.styled.ts
@@ -1,0 +1,3 @@
+import styled from "styled-components";
+
+export const Styledfinder = styled.div``;
