+++ src/components/pages/PokemonFinder/PokemonFinder.tsx
@@ -1,0 +1,3 @@
+export const PokemonFinder = () => {
+	return <span>Finder</span>;
+};
