+++ src/components/pages/PokemonFinder/index.ts
@@ -1,0 +1,1 @@
+export * from "./PokemonFinder";
