+++ src/main.tsx
@@ -1,9 +1,10 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import App from "@/components/pages/App";
+import App from "@/components/pages/Pokedex/Pokedex";
+import { Router } from "./Router";
 
 ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
 	<React.StrictMode>
-		<App />
+		<Router />
 	</React.StrictMode>
 );
